requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.2.1 # pyup: != 11.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.2  # https://github.com/ndparker/rcssmin
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.9.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==5.2  # pyup: < 5.2 # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-jet-reboot==1.3.10  # https://github.com/assem-ch/django-jet-reboot
django-guardian==3.0.0
# Django models utils
# ------------------------------------------------------------------------------
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-virtual-models==0.2.0
django-mptt==0.17.0
django-auditlog==3.1.2 # https://github.com/jazzband/django-auditlog

django-compressor==4.5.1  # https://github.com/django-compressor/django-compressor
django-filter==25.1   # https://github.com/carltongibson/django-filter
diskcache2==0.1.2 # https://github.com/grantjenks/python-diskcache